fix tests & update dependencies
update @homebridge/plugin-ui-utils
