+++ package.json
@@ -51,7 +51,7 @@
     "docs:theme": "typedoc --theme default-modern"
   },
   "dependencies": {
-    "@homebridge/plugin-ui-utils": "^2.0.0",
+    "@homebridge/plugin-ui-utils": "^2.0.1",
     "dorita980": "npm:@karlvr/dorita980@3.2.1"
   },
   "publishConfig": {
