+++ tsconfig.json
@@ -16,7 +16,7 @@
     "noImplicitAny": false,
     "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
   },
   "include": [
     "src"
